[derive] Refactor into more modules
- Move most top-level derive code into modules under a new top-level
  `derive` module
- Move some utilities into `util`
- Leave the root module to primarily perform routing and scaffolding

gherrit-pr-id: Gf40a4d73de475b2aae089e5e2d4cea53dd3c868c
